# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: "eBPF-SOC platform API"
  version: "1.0.0"
  description: |
    Two independent HTTP surfaces. They are NOT interchangeable and are
    documented together only because one console can drive both.

      * control-plane — multi-tenant. Every request is tenant-scoped and
        authorisation is evaluated per tenant. This is the MSSP/MSOC
        surface.
      * engine — single-tenant. One host, its own console, no tenancy
        model. This is the standalone product.

    A third surface, the agent-to-control-plane wire contract, is gRPC and
    is specified by the protobuf IDL in engine/proto/ebpfsoc/v1 — not here.

    Denials return 404, never 403, so a caller cannot use error codes to
    discover which tenants or resources exist. Treat a 404 as
    'no such thing, or not yours' — the distinction is deliberately
    unavailable.
servers:
  - url: "https://console.example.com"
    description: "Control plane (multi-tenant)"
  - url: "https://engine.example.com"
    description: "Single-tenant engine"
...
components:
  securitySchemes:
    bearerAdmin:
      type: http
      scheme: bearer
      description: |
        Control-plane admin token (CP_ADMIN_TOKEN). Grants the
        msoc-admin role across tenants. Machine-to-machine only.
    sessionCookie:
      type: apiKey
      in: cookie
      name: soc_session
      description: |
        Opaque session cookie. On the control plane it is issued by the
        BFF after an OIDC Authorization Code + PKCE flow, so an access
        token never reaches the browser. On the engine it is a signed,
        stateless cookie. Unsafe methods under /api/ additionally require
        the X-CSRF-Token header echoing the csrf_token cookie.
...
security:
  - sessionCookie: []
  - bearerAdmin: []
paths:
...
  "/api/login":
    post:
      tags: ["Authentication"]
      operationId: post_api_login
      summary: "HandleLogin"
      description: |
        HandleLogin parses form-encoded credentials and sets the session cookie.

        Surface: engine.
      responses:
        "200": { description: "Success" }
        "401": { description: "Unauthenticated" }
        "404": { description: "Not found, or not visible to this caller" }
  "/api/logout":
    get:
      tags: ["Authentication"]
      operationId: get_api_logout
      summary: "HandleLogout"
      description: |
        HandleLogout clears both cookies. With stateless sessions there's nothing
        server-side to invalidate; the cookies just stop being sent.

        Surface: engine.
      responses:
        "200": { description: "Success" }
        "401": { description: "Unauthenticated" }
        "404": { description: "Not found, or not visible to this caller" }
...
  "/favicon-light.svg":
    get:
      tags: ["Other"]
      operationId: get_favicon_light_svg
      summary: "handleFaviconLight"
      description: |
        No handler documentation in source.

        Surface: engine.
      responses:
        "200": { description: "Success" }
        "401": { description: "Unauthenticated" }
        "404": { description: "Not found, or not visible to this caller" }
  "/favicon.ico":
    get:
      tags: ["Other"]
      operationId: get_favicon_ico
      summary: "handleFaviconICO"
      description: |
        No handler documentation in source.

        Surface: engine.
      responses:
        "200": { description: "Success" }
        "401": { description: "Unauthenticated" }
        "404": { description: "Not found, or not visible to this caller" }
  "/favicon.svg":
    get:
      tags: ["Other"]
      operationId: get_favicon_svg
      summary: "handleFavicon"
      description: |
        No handler documentation in source.

        Surface: engine.
      responses:
        "200": { description: "Success" }
        "401": { description: "Unauthenticated" }
        "404": { description: "Not found, or not visible to this caller" }
...
  "/login":
    get:
      tags: ["Other"]
      operationId: get_login
      summary: "handleLoginPage"
      description: |
        No handler documentation in source.

        Surface: engine.
      responses:
        "200": { description: "Success" }
        "401": { description: "Unauthenticated" }
        "404": { description: "Not found, or not visible to this caller" }
...